requirements.txt
PyQt6==6.6.1
Pillow>=10.4.0
